requirements.txt
requests
python-dotenv
Flask
Flask-SQLAlchemy
Werkzeug
